+++ test/runtests.jl
@@ -1,5 +1,5 @@
 using BioSequenceGraphs
-using Base.Test
+using Test
 
 # write your own tests here
 @test 1 == 1
